# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Vitas Project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/expert": {
            "get": {
                "tags": [
                    "Experts"
                ],
                "summary": "Get all experts",
                "description": "Get all experts",
                "operationId": "da8f047daea1df0e6192b78e99992088",
                "responses": {
                    "200": {
                        "description": "Get Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Experts"
                ],
                "summary": "Create a new expert",
                "description": "Create a new expert",
                "operationId": "53ce1aee992a02ef09a193545953c5d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validated fail!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Something went wrong!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/expert/{id}": {
            "get": {
                "tags": [
                    "Experts"
                ],
                "summary": "Get expert by id",
                "description": "Get expert by id",
                "operationId": "847cc7299b53dbb135af1bb8c97f166b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the expert",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Experts"
                ],
                "summary": "Update a new expert",
                "description": "Update a new expert",
                "operationId": "2b4d086c4224cef9e17603b4f7ec85dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the expert",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validated fail!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Something went wrong!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Experts"
                ],
                "summary": "Delete expert by id",
                "description": "Delete expert by id",
                "operationId": "040217f2ec884f55a18a9f3397dfb207",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the expert",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/expert-detail": {
            "get": {
                "tags": [
                    "Expert detail"
                ],
                "summary": "Get all expert detail",
                "description": "Get all expert detail",
                "operationId": "3073e6c6e91488fe05bdd16443dafecf",
                "responses": {
                    "200": {
                        "description": "Get Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/expert-detail/{id}": {
            "get": {
                "tags": [
                    "Expert detail"
                ],
                "summary": "Get detail of expert by id",
                "description": "Get detail of expert by id",
                "operationId": "9de84c578fd47310af938d9846e3587a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the expert detail",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}